+++ stubs/SilverStripe/Core/ClassInfo.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace SilverStripe\Core;
+
+class ClassInfo
+{
+    public static function subclassesFor($class)
+    {
+    }
+
+    public static function getSubclassesFor($class)
+    {
+    }
+}